refactor: move addGame/; fix: loader/request on back from addGame
