+++ app/src/main/java/com/croumy/hltb_wearos/presentation/models/Constants.kt
@@ -18,7 +18,7 @@ class Constants {
         const val PREFERENCES_TOKEN = "hltb_token"
         const val PREFERENCES_USER_ID = "user_id"
 
-        const val HOME_NEED_REFRESH = "needRefresh"
+        const val HOME_NEED_REFRESH = "homeNeedRefresh"
         const val HOME_PREVIOUS_GAME_ID = "previousGameId"
         const val SEARCH_NEED_REFRESH = "searchNeedRefresh"
     }
