+++ app/src/main/java/com/croumy/hltb_wearos/presentation/navigation/NavActions.kt
@@ -1,10 +1,7 @@
 package com.croumy.hltb_wearos.presentation.navigation
 
 import androidx.navigation.NavHostController
-import androidx.navigation.NavOptions
-import androidx.navigation.Navigator
 import com.croumy.hltb_wearos.presentation.models.api.GameInfo
-import com.croumy.hltb_wearos.presentation.navigation.NavRoutes
 
 class NavActions(private val navController: NavHostController) {
     fun navigateBack() {
