+++ app/src/main/java/com/croumy/hltb_wearos/presentation/ui/addgame/AddGameViewModel.kt
@@ -1,4 +1,4 @@
-package com.croumy.hltb_wearos.presentation.ui.home.addgame
+package com.croumy.hltb_wearos.presentation.ui.addgame
 
 import android.content.Context
 import android.util.Log
@@ -16,7 +16,7 @@ import com.croumy.hltb_wearos.presentation.models.api.AddGameRequest
 import com.croumy.hltb_wearos.presentation.models.api.Category
 import com.croumy.hltb_wearos.presentation.models.api.GameInfo
 import com.croumy.hltb_wearos.presentation.models.api.QuickAdd
-import com.croumy.hltb_wearos.presentation.ui.home.addgame.models.AddGameStep
+import com.croumy.hltb_wearos.presentation.ui.addgame.models.AddGameStep
 import com.croumy.hltbwearos.R
 import dagger.hilt.android.lifecycle.HiltViewModel
 import dagger.hilt.android.qualifiers.ApplicationContext
