+++ app/src/main/java/com/croumy/hltb_wearos/presentation/ui/home/logs/LogsScreen.kt
@@ -22,6 +22,7 @@ import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Check
 import androidx.compose.material.icons.filled.Close
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.DisposableEffect
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
@@ -37,6 +38,8 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
+import androidx.lifecycle.Lifecycle
+import androidx.lifecycle.LifecycleEventObserver
 import androidx.lifecycle.lifecycleScope
 import androidx.wear.compose.foundation.lazy.ScalingLazyColumn
 import androidx.wear.compose.foundation.lazy.ScalingLazyListState
@@ -68,7 +71,7 @@ fun LogsScreen(
     LaunchedEffect(isFocusedScreen) {
         if (isFocusedScreen) {
             viewModel.getLogs()
-            if(viewModel.logs.value.isNotEmpty()) {
+            if (viewModel.logs.value.isNotEmpty()) {
                 // WAIT FOR THE ScalingLazyColumn TO BE READY
                 delay(100)
                 focusRequester.requestFocus()
